Refactor conjur.sdk.api -> conjur.sdk.endpoint
