+++ client/README.md
@@ -78,7 +78,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
