+++ client/docs/AuthenticationApi.md
@@ -35,7 +35,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -110,7 +110,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -190,7 +190,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -265,7 +265,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -342,7 +342,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -429,7 +429,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -518,7 +518,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -607,7 +607,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -691,7 +691,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -767,7 +767,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -856,7 +856,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -940,7 +940,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -1023,7 +1023,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.AuthenticationApi;
+import org.conjur.sdk.endpoint.AuthenticationApi;
 
 public class Example {
   public static void main(String[] args) {
