+++ client/docs/CertificateAuthorityApi.md
@@ -23,7 +23,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.CertificateAuthorityApi;
+import org.conjur.sdk.endpoint.CertificateAuthorityApi;
 
 public class Example {
   public static void main(String[] args) {
