+++ client/docs/HostFactoryApi.md
@@ -25,7 +25,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.HostFactoryApi;
+import org.conjur.sdk.endpoint.HostFactoryApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -100,7 +100,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.HostFactoryApi;
+import org.conjur.sdk.endpoint.HostFactoryApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -181,7 +181,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.HostFactoryApi;
+import org.conjur.sdk.endpoint.HostFactoryApi;
 
 public class Example {
   public static void main(String[] args) {
