+++ client/docs/PoliciesApi.md
@@ -25,7 +25,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.PoliciesApi;
+import org.conjur.sdk.endpoint.PoliciesApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -106,7 +106,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.PoliciesApi;
+import org.conjur.sdk.endpoint.PoliciesApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -187,7 +187,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.PoliciesApi;
+import org.conjur.sdk.endpoint.PoliciesApi;
 
 public class Example {
   public static void main(String[] args) {
