+++ client/docs/PublicKeysApi.md
@@ -23,7 +23,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.PublicKeysApi;
+import org.conjur.sdk.endpoint.PublicKeysApi;
 
 public class Example {
   public static void main(String[] args) {
