+++ client/docs/RolesApi.md
@@ -25,7 +25,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.RolesApi;
+import org.conjur.sdk.endpoint.RolesApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -108,7 +108,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.RolesApi;
+import org.conjur.sdk.endpoint.RolesApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -191,7 +191,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.RolesApi;
+import org.conjur.sdk.endpoint.RolesApi;
 
 public class Example {
   public static void main(String[] args) {
