+++ client/docs/StatusApi.md
@@ -26,7 +26,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.StatusApi;
+import org.conjur.sdk.endpoint.StatusApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -101,7 +101,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.StatusApi;
+import org.conjur.sdk.endpoint.StatusApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -177,7 +177,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.StatusApi;
+import org.conjur.sdk.endpoint.StatusApi;
 
 public class Example {
   public static void main(String[] args) {
@@ -258,7 +258,7 @@ import org.conjur.sdk.ApiException;
 import org.conjur.sdk.Configuration;
 import org.conjur.sdk.auth.*;
 import org.conjur.sdk.models.*;
-import org.conjur.sdk.api.StatusApi;
+import org.conjur.sdk.endpoint.StatusApi;
 
 public class Example {
   public static void main(String[] args) {
