+++ client/src/main/java/org/conjur/sdk/endpoint/StatusApi.java
@@ -11,7 +11,7 @@
  */
 
 
-package org.conjur.sdk.api;
+package org.conjur.sdk.endpoint;
 
 import org.conjur.sdk.ApiCallback;
 import org.conjur.sdk.ApiClient;
