+++ client/src/test/java/org/conjur/sdk/api/ConfiguredTest.java
@@ -1,4 +1,4 @@
-package org.conjur.sdk.api;
+package org.conjur.sdk.endpoint;
 
 import org.conjur.sdk.*;
 import org.conjur.sdk.auth.*;
