+++ config/java.yml
@@ -1,4 +1,4 @@
-apiPackage: org.conjur.sdk.api
+apiPackage: org.conjur.sdk.endpoint
 modelPackage: org.conjur.sdk.model
 scmUrl: https://github.com/cyberark/conjur-sdk-java
 invokerPackage: org.conjur.sdk
